+++ src/main/java/lermitage/intellij/extra/icons/ModelTag.java
@@ -13,7 +13,9 @@ public enum ModelTag {
     LIQUIBASE("Liquibase"),
     MS_OFFICE("MS Office"),
     NESTJS("NestJS"),
-    STORYBOOK("Storybook");
+    PRETTIER("Prettier"),
+    STORYBOOK("Storybook"),
+    TRAVIS("Travis CI");
 
     ModelTag(String name) {
         this.name = name;