+++ src/main/java/lermitage/intellij/extra/icons/cfg/dialogs/ModelDialog.java
@@ -15,8 +15,6 @@ import com.intellij.ui.ToolbarDecorator;
 import com.intellij.ui.components.JBLabel;
 import com.intellij.ui.components.JBTextField;
 import com.intellij.util.IconUtil;
-import lermitage.intellij.extra.icons.utils.AsyncUtils;
-import lermitage.intellij.extra.icons.utils.IconUtils;
 import lermitage.intellij.extra.icons.ExtraIconProvider;
 import lermitage.intellij.extra.icons.IconType;
 import lermitage.intellij.extra.icons.Model;
@@ -24,7 +22,8 @@ import lermitage.intellij.extra.icons.ModelCondition;
 import lermitage.intellij.extra.icons.ModelType;
 import lermitage.intellij.extra.icons.cfg.SettingsForm;
 import lermitage.intellij.extra.icons.cfg.SettingsService;
-import lermitage.intellij.extra.icons.providers.Angular2IconProvider;
+import lermitage.intellij.extra.icons.utils.AsyncUtils;
+import lermitage.intellij.extra.icons.utils.IconUtils;
 import org.jetbrains.annotations.Nullable;
 
 import javax.swing.Icon;
@@ -45,7 +44,6 @@ import java.awt.event.ItemEvent;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 import java.net.URI;
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Objects;
@@ -141,10 +139,8 @@ public class ModelDialog extends DialogWrapper {
         typeComboBox.addActionListener(e -> updateUIOnTypeChange());
 
         chooseIconSelector.addItem("choose custom or bundled icon");
-        List<Model> bundledModels = new ArrayList<>();
-        bundledModels.addAll(Angular2IconProvider.allModels());
-        bundledModels.addAll(ExtraIconProvider.allModels());
-        bundledModels.stream()
+        //bundledModels.addAll(Angular2IconProvider.allModels());
+        ExtraIconProvider.allModels().stream()
             .map(Model::getIcon)
             .sorted()
             .distinct()
@@ -247,7 +243,7 @@ public class ModelDialog extends DialogWrapper {
         if (model.getIconType() == IconType.PATH) {
             for (int itemIdx = 0; itemIdx < chooseIconSelector.getItemCount(); itemIdx++) {
                 Object item = chooseIconSelector.getItemAt(itemIdx);
-                if (item instanceof BundledIcon && ((BundledIcon)item).iconPath.equals(model.getIcon())) {
+                if (item instanceof BundledIcon && ((BundledIcon) item).iconPath.equals(model.getIcon())) {
                     chooseIconSelector.setSelectedIndex(itemIdx);
                     break;
                 }